+++ ports/freertos/timext.c
@@ -9,19 +9,19 @@
 #include "FreeRTOS.h"
 #include "task.h"
 
-#define time_after(goal, chasing)	((int)(goal)    - (int)(chasing) < 0)
+#define time_after(goal, chasing)	((long)(goal) - (long)(chasing) < 0)
 
-void timeout_set(unsigned int *goal, unsigned int msec)
+void timeout_set(unsigned long *goal, unsigned long msec)
 {
 	*goal = xTaskGetTickCount() + pdMS_TO_TICKS(msec);
 }
 
-bool timeout_is_expired(unsigned int goal)
+bool timeout_is_expired(unsigned long goal)
 {
 	return time_after(goal, xTaskGetTickCount());
 }
 
-void sleep_ms(unsigned int msec)
+void sleep_ms(unsigned long msec)
 {
 	vTaskDelay(pdMS_TO_TICKS(msec));
 }
